C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik5.cpp -o C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik5.exe  -Wshadow -Winit-self -Wredundant-decls -Wcast-align -Wundef -Wfloat-equal -Winline -Wunreachable-code -Wmissing-declarations -Wmissing-include-dirs -Wswitch-enum -Wswitch-default -Weffc++ -Wmain -Wextra -Wall -g -pipe -fexceptions -Wcast-qual -Wconversion -Wctor-dtor-privacy -Wempty-body -Wformat-security -Wformat=2 -Wignored-qualifiers -Wlogical-op -Wmissing-field-initializers -Wnon-virtual-dtor -Woverloaded-virtual -Wpointer-arith -Wsign-promo -Wstack-usage=8192 -Wstrict-aliasing -Wstrict-null-sentinel -Wtype-limits -Wwrite-strings -D_DEBUG -D_EJUDGE_CLIENT_SIDE -D_EJC

+++ Ozhaev Zaynudin Sharik5.cpp
@@ -23,19 +23,19 @@ int main()
 void MoveBall()
     {
     srand(time(NULL));
-    float  x1 = rand() % W,   y1 = rand() % H,
+      int  x1 = rand() % W,   y1 = rand() % H,
           vx1 = rand() % 10, vy1 = rand() % 10,
            r1 = 15 + rand() % 35;
 
-    float x2 = rand() % W,   y2 = rand() % H,
+      int x2 = rand() % W,   y2 = rand() % H,
          vx2 = rand() % 10, vy2 = rand() % 10,
           r2 = 15 + rand() % 35;
 
-    float x3 = rand() % W,   y3 = rand() % H,
+      int x3 = rand() % W,   y3 = rand() % H,
          vx3 = rand() % 10, vy3 = rand() % 10,
           r3 = 15 + rand() % 35;
 
-    float x4 = rand() % W,   y4 = rand() % H,
+      int x4 = rand() % W,   y4 = rand() % H,
          vx4 = rand() % 10, vy4 = rand() % 10,
           r4 = 15 + rand() % 35;
